+++ sequana/datatools.py
@@ -20,22 +20,58 @@
 import os
 import easydev
 import glob
+import collections
 
 
 def sequana_data(filename=None, where=None):
-    """Simple utilities to retrieve data sets from sequana resources
+    """Return full path of a sequana resource data file.
+
+    :param str filename: a valid filename to be found
+    :param str where: one of the registered data directory (see below)
+    :return: the path of file. See also here below in the case where 
+        filename is set to "*".
 
     .. code-block:: python
 
         from sequana import sequana_data
         filename = sequana_data("test.bam")
 
-    Type the function name without parameter to get a list of available files
+    Type the function name with "*" parameter to get a list of
+    available files. Withe where argument set, the function returns a 
+    list of files. Without the where argument, a dictionary is returned where
+    keys correspond to the registered directories::
+
+        filenames = sequana_data("*", where="images")
+
+    Registered directories are:
+
+        - data
+        - testing
+        - data/adapters
+        - images
+
+    .. note:: this does not handle wildcards. The * means retrieve all files.
 
     """
     sequana_path = easydev.get_package_location('sequana')
     sharedir = os.sep.join([sequana_path , "sequana", 'resources'])
     directories = ['data', 'testing', 'data/adapters', 'images']
+
+    if filename == "*":
+        found = collections.defaultdict(list)
+        if where is not None:
+            directories = [where]
+        for thisdir in directories:
+            for filename in glob.glob(sharedir + "/%s/*" % thisdir):
+                filename = os.path.split(filename)[1]
+                to_ignore = ["__init__.py", "__pycache__"]
+                if filename.endswith('.pyc') or filename in to_ignore:
+                    pass
+                else:
+                    found[thisdir].append(os.path.split(filename)[1])
+        if where is not None: 
+            return found[where]
+        return found
 
     if filename is None:
         for thisdir in directories:
@@ -48,6 +84,7 @@ def sequana_data(filename=None, where=None):
                 else:
                     print(' - sequana("%s", "%s")' % (os.path.split(filename)[1], thisdir))
         raise ValueError("Choose a valid file from the list above")
+
     # in the code one may use / or \ 
     if where:
         filename = os.sep.join([sharedir, where, filename])