+++ sequana/errors.py
@@ -1,11 +1,5 @@
 
 
-
-
-
-
-
-
 class SequanaException(Exception):
     def __init__(self, *args, **kwargs):
         Exception.__init__(self, *args, **kwargs)